requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    bazel run //:python_deps.update
#
absl-py==2.1.0 \
    --hash=sha256:526a04eadab8b4ee719ce68f204172ead1027549089702d99b9059f129ff1308 \
    --hash=sha256:7820790efbb316739cde8b4e19357243fc3608a152024288513dd968d7d959ff
    # via -r requirements.in
